+++ include/graphics/DirectionField.h
@@ -11,10 +11,10 @@ public:
 public:
 	static void			init(int layerNb);
 	static void         initVectorField(const std::vector<std::vector<Shape>> &zones, const std::vector<Shape> &borders, int layerIndex);
-	static glm::vec2    getVecAtPos(const glm::vec2 &point, int layerIndex);
-	static glm::vec2    getVecUnder(const glm::vec2 &a, const glm::vec2 &b, int layerIndex);
+	static glm::dvec2   getVecAtPos(const glm::dvec2 &point, int layerIndex);
+	static glm::dvec2   getVecUnder(const glm::dvec2 &a, const glm::dvec2 &b, int layerIndex);
 
 private:
-	static inline std::vector<std::vector<glm::vec2>>   _vectorField;
-	static inline std::vector<std::vector<glm::vec2>>   _vectorFieldSum;
+	static inline std::vector<std::vector<glm::dvec2>>   _vectorField;
+	static inline std::vector<std::vector<glm::dvec2>>   _vectorFieldSum;
 };
