+++ include/localAlgo/LocalOperator.h
@@ -16,7 +16,7 @@ private:
 	unsigned int												_iter = 0;
 	bool														_isEnd = false;
 	std::vector<int>											_index;
-	std::pair<std::vector<glm::vec3>, std::vector<glm::vec3>>	_final;
+	std::pair<std::vector<glm::dvec3>, std::vector<glm::dvec3>>	_final;
 	std::vector<Shape>											_colorZones;
 	std::mt19937												_gen;
 
@@ -28,19 +28,19 @@ public:
 	void	updateState(bool);
 	void	startShuffling(int, int);
 	void	optimize();
-	const std::pair<std::vector<glm::vec3>, std::vector<glm::vec3>>&	getFinal() const;
+	const std::pair<std::vector<glm::dvec3>, std::vector<glm::dvec3>>&	getFinal() const;
 	const std::vector<Shape>&											getColorZones() const;
 
 private:
 	bool	checkPaperOp(int);
 	bool	isLinked(const std::vector<int> &, int);
-	bool	switchConditions(float);
+	bool	switchConditions(double);
 	void	computeIndex(int, int);
 
-	std::pair<float, std::vector<int>> 	checkFlip(const std::vector<int> &);
-	std::pair<float, std::vector<int>> 	checkTranspose(const std::vector<int> &);
-	std::pair<float, std::vector<int>> 	checkCross(const std::vector<int> &);
-	std::pair<float, std::vector<int>> 	checkZigZag(const std::vector<int> &);
+	std::pair<double, std::vector<int>> 	checkFlip(const std::vector<int> &);
+	std::pair<double, std::vector<int>> 	checkTranspose(const std::vector<int> &);
+	std::pair<double, std::vector<int>> 	checkCross(const std::vector<int> &);
+	std::pair<double, std::vector<int>> 	checkZigZag(const std::vector<int> &);
 
 };
